HYD-5579 Big Differ: Refactor to components
HYD-5005 has been difficult to implement. Partly because we keep lot's
of state in component controllers when using big-differ.

We can solve this problem by refactoring big-differ to be a series of
components instead. The approach is this:

1) Wire into ngModelCtrl $parsers and $formatters. This will give us a
way to update or decide not to update model / view values.

2) Create a component responsible for managing diff states. This will
give us a centralized place to bookeep a input components diff state.

3) Create an overall component that is able to manage multiple child
diff components. We have some controls like Configure LNet where we have
multiple diff selects. By having a component the child diffs can
require, we can do bulk dirty checks and reset actions.

Once these components are implemented, we can remove all diff state from
controllers and delete the old big-differ implementation.

This should be done in an external module so it can be easily backported
between branches.

  - Create initial repo, code and tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>